wip-upgrade: use hooks and the new toys, refactor
