+++ src/components/common/button.tsx
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const Button = styled.button`
+  background: #444;
+  border-radius: 3px;
+  padding: 1rem;
+  border: none;
+  height: 68px;
+  width: 100%;
+  cursor: pointer;
+  box-shadow: -2px 2px 2px rgba(0, 0, 0, 0.4);
+`;
