+++ src/types.d.ts
@@ -1,0 +1,11 @@
+type Kit = {
+  id: number;
+  name: string;
+  sounds: {
+    id: number;
+    name: string;
+    src: string;
+    keyCode: number;
+    keyChar: string;
+  }[];
+};
